requirements.txt
attrs==21.4.0
automata-lib==5.0.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydot==1.4.2
pyparsing==3.0.9
pytest==7.1.2
tomli==2.0.1
